+++ build.sh
@@ -15,4 +15,6 @@ fi
 
 rm -f *.mod
 
-"${GFC}" -Wall -Wextra -pedantic -std=f2018 -O3 src/forsynth.f90 src/envelopes.f90 src/signals.f90 src/music.f90 src/audio_effects.f90 src/demos.f90 app/main.f90 -o build/forsynth.out
+"${GFC}" -Wall -Wextra -pedantic -std=f2018 -O3 src/forsynth.f90 src/envelopes.f90 src/signals.f90 src/music.f90 src/audio_effects.f90 example/demo1.f90 -o build/demo1.out
+"${GFC}" -Wall -Wextra -pedantic -std=f2018 -O3 src/forsynth.f90 src/envelopes.f90 src/signals.f90 src/music.f90 src/audio_effects.f90 example/demo2.f90 -o build/demo2.out
+"${GFC}" -Wall -Wextra -pedantic -std=f2018 -O3 src/forsynth.f90 src/envelopes.f90 src/signals.f90 src/music.f90 src/audio_effects.f90 example/demo3.f90 -o build/demo3.out