+++ test/main.f90
@@ -1,10 +1,13 @@
 ! Forsynth: a multitracks stereo sound synthesis project
 ! License GPL-3.0-or-later
 ! Vincent Magnin
-! Last modification: 2021-04-13
+! Last modification: 2024-04-20
 
 program main
-implicit none
+    use forsynth, only: test_the_machine
 
-print *, "Put some tests in here!"
+    implicit none
+
+    call test_the_machine()
+
 end program main